backend/pyproject.toml
[project]
name = "zora"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "helium>=5.1.1",
    "matplotlib>=3.10.1",
    "python-dotenv>=1.1.0",
    "selenium>=4.31.0",
    "smolagents[openai]>=1.13.0",
]
